requirements.txt
# this is where I will put any packages that I need to download into my conda environment
pandas
python-dotenv
sendgrid==6.0.5

Flask

#heroku needs this to run our app on the server
gunicorn